feat: enable secureboot

+++ scripts/chroot_build.sh
@@ -180,14 +180,11 @@ function build_image() {
     cp /boot/initrd.img-**-**-generic casper/initrd
 
     # memtest86
-    case $TARGET_UBUNTU_VERSION in
-        "noble")
-            cp /boot/memtest86+x64.bin install/memtest86+
-            ;;
-        *)
-            cp /boot/memtest86+.bin install/memtest86+
-            ;;
-    esac
+    if [ -f "/boot/memtest86+x64.bin" ]; then
+        cp /boot/memtest86+x64.bin install/memtest86+
+    else
+        cp /boot/memtest86+.bin install/memtest86+
+    fi
 
     # memtest86++
     wget --progress=dot https://www.memtest86.com/downloads/memtest86-usb.zip -O install/memtest86-usb.zip
